Added OpenTelemetry support
- Added extension/OpenTelemetry.js to automatically wrap exec / execute /
prepare / commit / rollback with OpenTelemetry tracing
    - Client will automatically wrap with OpenTelemetry tracing when
      isOpenTelemetryEnabled() is true
    - Most async functions will propagate context (except connect to match
      hana-client)
- Added openTelemetry and openTelemetryTenant connect options
- Added isOpenTelemetrySupported property to the driver (Hana class)
- Added MockOpenTelemetrySDK and integration tests

Testing Notes
OpenTelemetry is only tested when the ENABLE_NODE_OPENTEL_TESTS environment
variable is set and not 0 or false. So the Makefile now has `make test-opentelemetry`
which will run the OpenTelemetry integration tests with this environment variable.

To install the MockOpenTelemetrySDK, copy the contents of the MockOpenTelemetrySDK
into node_modules/@opentelemetry/api